Separate hoist section of analyzer
